[RORDEV-1636] Give the ES|QL API a suite of its own, run in both X-Pack modes
The ES|QL cases sat in BaseXpackApiSuite next to everything else X-Pack, and
LOOKUP JOIN was covered twice - there and in a standalone suite with its own
cluster. They now live in BaseEsqlApiSuite, whose cluster holds only what ES|QL
reads, and run against both an X-Pack-secured and a ROR-secured node.
