+++ integration-tests/src/test/resources/esql_api/readonlyrest_without_ror_ssl.yml
@@ -1,0 +1,49 @@
+readonlyrest:
+
+  access_control_rules:
+
+    # ES container initializer need this rule to configure ES instance after startup
+    - name: "CONTAINER ADMIN"
+      type: allow
+      auth_key: admin:container
+
+    - name: "admin esql"
+      verbosity: error
+      type: allow
+      indices: ["bookstore", "library", "bookshop"]
+      auth_key: sqladmin:pass
+
+    - name: "dev1 esql"
+      auth_key: dev1sql:test
+      indices: ["bookstore"]
+      fields: ["~price"]
+
+    - name: "dev2 esql"
+      auth_key: dev2sql:test
+      indices: ["library"]
+      fields: ["~internal_id"]
+
+    - name: "dev3 esql"
+      auth_key: dev3sql:test
+      indices: ["bookstore"]
+      filter: '{"query_string":{"query":"name:(Dune)"}}'
+
+    - name: "Catalog only - the LOOKUP JOIN target is out of reach"
+      auth_key: catalog:test
+      indices: ["book_catalog"]
+
+    - name: "Prices only - the FROM target is out of reach"
+      auth_key: prices:test
+      indices: ["book_prices"]
+
+    - name: "Both the FROM and the LOOKUP JOIN target"
+      auth_key: both:test
+      indices: ["book_catalog", "book_prices"]
+
+    - name: "Everything matching the book_ prefix"
+      auth_key: books:test
+      indices: ["book_*"]
+
+    - name: "The FROM target and both LOOKUP JOIN targets"
+      auth_key: joins:test
+      indices: ["book_catalog", "book_prices", "store_ratings"]
