+++ integration-tests/src/test/scala/tech/beshu/ror/integration/suites/base/BaseXpackApiSuite.scala
@@ -27,7 +27,6 @@ import tech.beshu.ror.utils.containers.*
 import tech.beshu.ror.utils.elasticsearch.*
 import tech.beshu.ror.utils.httpclient.RestClient
 import tech.beshu.ror.utils.misc.CustomScalaTestMatchers
-import tech.beshu.ror.utils.misc.Version
 
 trait BaseXpackApiSuite
     extends AnyWordSpecLike
@@ -72,15 +71,6 @@ trait BaseXpackApiSuite
   private lazy val dev1SqlManager = new SqlApiManager(basicAuthClient("dev1sql", "test"), esVersionUsed)
   private lazy val dev2SqlManager = new SqlApiManager(basicAuthClient("dev2sql", "test"), esVersionUsed)
   private lazy val dev3SqlManager = new SqlApiManager(basicAuthClient("dev3sql", "test"), esVersionUsed)
-
-  private lazy val adminEsqlManager = new EsqlApiManager(basicAuthClient("sqladmin", "pass"), esVersionUsed)
-  private lazy val dev1EsqlManager = new EsqlApiManager(basicAuthClient("dev1sql", "test"), esVersionUsed)
-  private lazy val dev2EsqlManager = new EsqlApiManager(basicAuthClient("dev2sql", "test"), esVersionUsed)
-  private lazy val dev3EsqlManager = new EsqlApiManager(basicAuthClient("dev3sql", "test"), esVersionUsed)
-  private lazy val dev4EsqlManager = new EsqlApiManager(basicAuthClient("dev4sql", "test"), esVersionUsed)
-  private lazy val dev5EsqlManager = new EsqlApiManager(basicAuthClient("dev5sql", "test"), esVersionUsed)
-  private lazy val dev6EsqlManager = new EsqlApiManager(basicAuthClient("dev6sql", "test"), esVersionUsed)
-  private lazy val dev7EsqlManager = new EsqlApiManager(basicAuthClient("dev7sql", "test"), esVersionUsed)
 
   "Async search" should {
     "be allowed for dev1 and test1_index_a" excludeES (allEs6x, allEs7xBelowEs77x) in {
@@ -1085,289 +1075,6 @@ trait BaseXpackApiSuite
     }
   }
 
-  "ESQL query request" when {
-    "SELECT query is used" should {
-      "be allowed" when {
-        "user has no indices rule (has access to any index)" when {
-          "full index name is used" excludeES (allEs6x, allEs7x, allEs8xBelowEs811x) in {
-            val result = adminEsqlManager.execute("""FROM library | LIMIT 100""")
-            result should have statusCode 200
-            result.columnNames should contain only (
-              "author",
-              "author.keyword",
-              "internal_id",
-              "name",
-              "name.keyword",
-              "release_date"
-            )
-            result.column("author").toList should contain only (Str("James S.A. Corey"), Str("Dan Simmons"))
-            result.column("internal_id").toList should contain only (Num(1), Num(2))
-            result.rows.size should be(2)
-          }
-          "full indices names are used" excludeES (allEs6x, allEs7x, allEs8xBelowEs811x) in {
-            val result = adminEsqlManager.execute("""FROM bookstore,library | LIMIT 100""")
-            result should have statusCode 200
-            result.columnNames should contain only (
-              "author",
-              "author.keyword",
-              "internal_id",
-              "name",
-              "name.keyword",
-              "release_date",
-              "price"
-            )
-            result.column("author").toList should contain only (
-              Str("James S.A. Corey"),
-              Str("Dan Simmons"),
-              Str("Frank Herbert")
-            )
-            result.column("internal_id").toList should contain only (Num(1), Num(2), Null)
-            result.rows.size should be(5)
-          }
-          "wildcard is used" excludeES (allEs6x, allEs7x, allEs8xBelowEs811x) in {
-            val result = adminEsqlManager.execute("""FROM * | LIMIT 100""")
-            result should have statusCode 200
-            result.columnNames should contain only (
-              "author",
-              "author.keyword",
-              "internal_id",
-              "name",
-              "name.keyword",
-              "release_date",
-              "price"
-            )
-            result.column("author").toList should contain only (
-              Str("James S.A. Corey"),
-              Str("Dan Simmons"),
-              Str("Frank Herbert")
-            )
-            result.column("internal_id").toList should contain only (Num(1), Num(2), Null)
-            result.rows.size should be(5)
-          }
-          "alias is used" excludeES (allEs6x, allEs7x, allEs8xBelowEs811x) in {
-            val result = adminEsqlManager.execute("""FROM bookshop | LIMIT 100""")
-            result should have statusCode 200
-            result.columnNames should contain only (
-              "author",
-              "author.keyword",
-              "name",
-              "name.keyword",
-              "price",
-              "release_date"
-            )
-            result.column("author").toList should contain only (
-              Str("James S.A. Corey"),
-              Str("Dan Simmons"),
-              Str("Frank Herbert")
-            )
-            result.column("price").toList should contain only (Num(100), Num(200), Num(50))
-            result.rows.size should be(3)
-          }
-        }
-        "user has access to given index" when {
-          "full index name is used" excludeES (allEs6x, allEs7x, allEs8xBelowEs811x) in {
-            val result = dev1EsqlManager.execute("""FROM bookstore | LIMIT 100""")
-            result should have statusCode 200
-            result.columnNames should contain only ("author", "author.keyword", "name", "name.keyword", "release_date")
-            result.column("author").toList should contain only (
-              Str("James S.A. Corey"),
-              Str("Dan Simmons"),
-              Str("Frank Herbert")
-            )
-            result.rows.size should be(3)
-          }
-          "full indices names are used and one of them is not allowed" excludeES (
-            allEs6x,
-            allEs7x,
-            allEs8xBelowEs811x
-          ) in {
-            val result = dev1EsqlManager.execute("""FROM bookstore,library | LIMIT 100""")
-            result should have statusCode 200
-            result.columnNames should contain only ("author", "author.keyword", "name", "name.keyword", "release_date")
-            result.column("author").toList should contain only (
-              Str("James S.A. Corey"),
-              Str("Dan Simmons"),
-              Str("Frank Herbert")
-            )
-            result.rows.size should be(3)
-          }
-          "full indices names are used, one of them is not allowed and they are separated by a space" excludeES (
-            allEs6x,
-            allEs7x,
-            allEs8xBelowEs811x
-          ) in {
-            val result = dev1EsqlManager.execute("""FROM bookstore, library | LIMIT 100""")
-            result should have statusCode 200
-            result.columnNames should contain only ("author", "author.keyword", "name", "name.keyword", "release_date")
-            result.column("author").toList should contain only (
-              Str("James S.A. Corey"),
-              Str("Dan Simmons"),
-              Str("Frank Herbert")
-            )
-            result.rows.size should be(3)
-          }
-          "wildcard is used" excludeES (allEs6x, allEs7x, allEs8xBelowEs811x) in {
-            val result = dev1EsqlManager.execute("""FROM book* | LIMIT 100""")
-            result should have statusCode 200
-            result.columnNames should contain only ("author", "author.keyword", "name", "name.keyword", "release_date")
-            result.column("author").toList should contain only (
-              Str("James S.A. Corey"),
-              Str("Dan Simmons"),
-              Str("Frank Herbert")
-            )
-            result.rows.size should be(3)
-          }
-          "alias is used" excludeES (allEs6x, allEs7x, allEs8xBelowEs811x) in {
-            val result = dev1EsqlManager.execute("""FROM bookshop | LIMIT 100""")
-            result should have statusCode 200
-            result.columnNames should contain only ("author", "author.keyword", "name", "name.keyword", "release_date")
-            result.column("author").toList should contain only (
-              Str("James S.A. Corey"),
-              Str("Dan Simmons"),
-              Str("Frank Herbert")
-            )
-            result.rows.size should be(3)
-          }
-          "filter in block is used" excludeES (allEs6x, allEs7x, allEs8xBelowEs811x) in {
-            val result = dev3EsqlManager.execute("""FROM bookstore | LIMIT 100""")
-            result should have statusCode 200
-            result.columnNames should contain only (
-              "author",
-              "author.keyword",
-              "name",
-              "name.keyword",
-              "price",
-              "release_date"
-            )
-            result.column("author").toList should contain only Str("Frank Herbert")
-            result.rows.size should be(1)
-          }
-        }
-        "ESQL keywords are not uppercased" excludeES (allEs6x, allEs7x, allEs8xBelowEs811x) in {
-          val result = dev1EsqlManager.execute("""fRoM book* | lImIt 100""")
-          result should have statusCode 200
-          result.columnNames should contain only ("author", "author.keyword", "name", "name.keyword", "release_date")
-          result.column("author").toList should contain only (
-            Str("James S.A. Corey"),
-            Str("Dan Simmons"),
-            Str("Frank Herbert")
-          )
-          result.rows.size should be(3)
-        }
-      }
-      "be bad request (implicitly forbidden)" when {
-        "user doesn't have access to given index" when {
-          "full index name is used" excludeES (allEs6x, allEs7x, allEs8xBelowEs811x) in {
-            val result = dev2EsqlManager.execute("""FROM bookstore | LIMIT 100""")
-            result should have statusCode 400
-            result.responseJson("error").obj("reason").str should include("Unknown index")
-          }
-          "wildcard is used" excludeES (allEs6x, allEs7x, allEs8xBelowEs811x, allEs9xAboveEs93x) in {
-            val result = dev2EsqlManager.execute("""FROM book* | LIMIT 100""")
-            result should have statusCode 400
-            result.responseJson("error").obj("reason").str should include("Unknown index")
-          }
-          "alias is used" excludeES (allEs6x, allEs7x, allEs8xBelowEs811x) in {
-            val result = dev2EsqlManager.execute("""FROM bookshop | LIMIT 100""")
-            result should have statusCode 400
-            result.responseJson("error").obj("reason").str should include("Unknown index")
-          }
-          "not-existent index name is used" excludeES (allEs6x, allEs7x, allEs8xBelowEs811x) in {
-            val result = dev2EsqlManager.execute("""FROM flea_market | LIMIT 100""")
-            result should have statusCode 400
-            result.responseJson("error").obj("reason").str should include("Unknown index")
-          }
-        }
-        "user rule is not used" when {
-          "not-existent index name is used" excludeES (allEs6x, allEs7x, allEs8xBelowEs811x) in {
-            val result = adminEsqlManager.execute("""FROM unknown | LIMIT 100""")
-            result should have statusCode 400
-          }
-        }
-        "esql query is malformed" excludeES (allEs6x, allEs7x, allEs8xBelowEs811x) in {
-          val result = adminEsqlManager.execute("""FROM unescaped-index.name | LIMIT 100""")
-          result should have statusCode 400
-        }
-      }
-      "return empty result (implicitly forbidden)" when {
-        "user doesn't have access to given index" when {
-          "wildcard is used" excludeES (allEs6x, allEs7x, allEs8x, allEs9xBelowEs94x) in {
-            val result = dev2EsqlManager.execute("""FROM book* | LIMIT 100""")
-            result should have statusCode 200
-            result.rows should be(empty)
-          }
-        }
-      }
-    }
-    "LOOKUP JOIN is used" should {
-      "be allowed" when {
-        "the LOOKUP JOIN target is authorized" excludeES (allEs6x, allEs7x, allEs8xBelowEs818x) in {
-          val result = dev7EsqlManager.execute(
-            """FROM book_catalog | LOOKUP JOIN book_prices ON book_id | SORT book_id | LIMIT 100"""
-          )
-          result should have statusCode 200
-          result.columnNames should contain only ("book_id", "title", "title.keyword", "discount_price")
-          result.column("discount_price").toList should contain only (Num(90), Num(180))
-        }
-        "the narrowed FROM wildcard also matches the LOOKUP JOIN target" excludeES (
-          allEs6x,
-          allEs7x,
-          allEs8xBelowEs818x
-        ) in {
-          val result = dev5EsqlManager.execute(
-            """FROM book_* | LOOKUP JOIN book_prices ON book_id | SORT book_id | LIMIT 100"""
-          )
-          result should have statusCode 200
-          result.columnNames should contain only ("book_id", "title", "title.keyword", "discount_price")
-          result.column("title").toList should contain only (Str("Leviathan Wakes"), Str("Hyperion"), Null)
-          result.rows.size should be(4)
-        }
-      }
-      "narrow a comma-separated FROM list to what the user is granted" when {
-        "the entries are separated by a space" excludeES (allEs6x, allEs7x, allEs8xBelowEs818x) in {
-          val result = dev4EsqlManager.execute("""FROM book_catalog, book_prices | SORT book_id | LIMIT 100""")
-          result should have statusCode 200
-          result.columnNames should contain only ("book_id", "title", "title.keyword")
-          result.rows.size should be(2)
-        }
-        "the entries are quoted" excludeES (allEs6x, allEs7x, allEs8xBelowEs818x) in {
-          val result = dev4EsqlManager.execute("""FROM \"book_catalog\",\"book_prices\" | SORT book_id | LIMIT 100""")
-          result should have statusCode 200
-          result.columnNames should contain only ("book_id", "title", "title.keyword")
-          result.rows.size should be(2)
-        }
-      }
-      "deny the whole request with a generic 'Unknown index' error" when {
-        "the LOOKUP JOIN target is not authorized, even though FROM's own target is fine on its own" excludeES (
-          allEs6x,
-          allEs7x,
-          allEs8xBelowEs818x
-        ) in {
-          val result = dev4EsqlManager.execute(
-            """FROM book_catalog | LOOKUP JOIN book_prices ON book_id | LIMIT 100"""
-          )
-          result should have statusCode 400
-          val reason = result.responseJson("error").obj("reason").str
-          reason should include("Unknown index")
-          reason should not include "book_prices"
-        }
-        "FROM's own target is not authorized, even though the LOOKUP JOIN target is fine on its own" excludeES (
-          allEs6x,
-          allEs7x,
-          allEs8xBelowEs818x
-        ) in {
-          val result = dev6EsqlManager.execute(
-            """FROM book_catalog | LOOKUP JOIN book_prices ON book_id | LIMIT 100"""
-          )
-          result should have statusCode 400
-          val reason = result.responseJson("error").obj("reason").str
-          reason should include("Unknown index")
-          reason should not include "book_catalog"
-        }
-      }
-    }
-  }
-
   "Get terms request" should {
     "be allowed for dev1 and test1_index_a" excludeES (allEs6x, allEs7xBelowEs714x) in {
       val result = dev1XpackApiManager.getTerms("test1_index_a", "hello.keyword")
@@ -1423,9 +1130,6 @@ object BaseXpackApiSuite {
     storeScriptTemplate(adminRestClient, esVersion)
     configureBookstore(documentManager, indexManager)
     configureLibrary(documentManager)
-    if (Version.greaterOrEqualThan(esVersion, 8, 18, 0)) {
-      configureBookPrices(documentManager, indexManager)
-    }
 
     indexManager.closeIndex("test3_index_c").force()
   }
@@ -1538,40 +1242,6 @@ object BaseXpackApiSuite {
       )
     )
     indexManager.createAliasOf("bookstore", "bookshop").force()
-  }
-
-  private def configureBookPrices(documentManager: DocumentManager, indexManager: IndexManager): Unit = {
-    documentManager.createDocAndAssert(
-      index = "book_catalog",
-      `type` = "_doc",
-      id = 1,
-      content = ujson.read("""{"book_id": 1, "title": "Leviathan Wakes"}""")
-    )
-    documentManager.createDocAndAssert(
-      index = "book_catalog",
-      `type` = "_doc",
-      id = 2,
-      content = ujson.read("""{"book_id": 2, "title": "Hyperion"}""")
-    )
-
-    indexManager
-      .createIndex(
-        "book_prices",
-        settings = Some(ujson.read("""{"settings": {"index": {"mode": "lookup"}}}"""))
-      )
-      .force()
-    documentManager.createDocAndAssert(
-      index = "book_prices",
-      `type` = "_doc",
-      id = 1,
-      content = ujson.read("""{"book_id": 1, "discount_price": 90}""")
-    )
-    documentManager.createDocAndAssert(
-      index = "book_prices",
-      `type` = "_doc",
-      id = 2,
-      content = ujson.read("""{"book_id": 2, "discount_price": 180}""")
-    )
   }
 
   private def configureLibrary(documentManager: DocumentManager): Unit = {
